♻️ Refactor: 인증 화면 MVI Contract 구조 변경
- `certification`, `detail`, `editor` 각 화면의 MVI (Model-View-Intent) 관련 클래스들을 `contact` 패키지로 이동하여 구조를 통일합니다.
- 패키지 구조 변경에 따라 관련 import 구문을 수정합니다.
