+++ feature/task-certification/src/main/java/com/twix/task_certification/certification/contract/TaskCertificationUiState.kt
@@ -1,9 +1,12 @@
-package com.twix.task_certification.certification.model
+package com.twix.task_certification.certification.contract
 
 import android.net.Uri
 import androidx.camera.core.CameraSelector
 import androidx.compose.runtime.Immutable
 import com.twix.designsystem.components.comment.model.CommentUiModel
+import com.twix.task_certification.certification.model.CameraPreview
+import com.twix.task_certification.certification.model.CaptureStatus
+import com.twix.task_certification.certification.model.TorchStatus
 import com.twix.ui.base.State
 
 @Immutable
@@ -35,7 +38,7 @@ data class TaskCertificationUiState(
     }
 
     fun toggleTorch(): TaskCertificationUiState {
-        val newFlashMode = TorchStatus.toggle(torch)
+        val newFlashMode = TorchStatus.Companion.toggle(torch)
         return copy(torch = newFlashMode)
     }
