+++ feature/task-certification/src/main/java/com/twix/task_certification/detail/contract/TaskCertificationDetailSideEffect.kt
@@ -1,11 +1,11 @@
-package com.twix.task_certification.editor.model
+package com.twix.task_certification.detail.contract
 
 import com.twix.designsystem.components.toast.model.ToastType
 import com.twix.ui.base.SideEffect
 
-sealed interface TaskCertificationEditorSideEffect : SideEffect {
+sealed interface TaskCertificationDetailSideEffect : SideEffect {
     data class ShowToast(
         val message: Int,
         val type: ToastType,
-    ) : TaskCertificationEditorSideEffect
+    ) : TaskCertificationDetailSideEffect
 }
