+++ feature/task-certification/src/main/java/com/twix/task_certification/editor/contract/TaskCertificationEditorSideEffect.kt
@@ -1,11 +1,11 @@
-package com.twix.task_certification.detail.model
+package com.twix.task_certification.editor.contract
 
 import com.twix.designsystem.components.toast.model.ToastType
 import com.twix.ui.base.SideEffect
 
-sealed interface TaskCertificationDetailSideEffect : SideEffect {
+sealed interface TaskCertificationEditorSideEffect : SideEffect {
     data class ShowToast(
         val message: Int,
         val type: ToastType,
-    ) : TaskCertificationDetailSideEffect
+    ) : TaskCertificationEditorSideEffect
 }
